+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 ### Fixed
 - #ifdef directive lexing bug
 - Macro define already exists threw an error
-- Macro function call with param that are themselves macros threw an error 
+- Macro function call with param that are themselves macros threw an error
+- Builtin shader variables coloring didn't work
 
 ## [1.1.0]
 ### Updated